// client/src/types/types.ts
export type DepartmentData = {
    id: number;
    name: string;
    vision: string;
    mission: string;
    organization: string;
    head: string;
  };

export type CurriculumData = {
    Sl_No: any;
    curriculum: string;
    program: string;
    Department: string;
    From: string;
    To: string;
    viewdetails: string;
  };